Add try_ macro for try/catch functionality
try/catch is a construct to catch errors in a local block and to handle
errors or exceptions that might occur. In python this might look like

```
try:
    val = some_dict["key_does_not_exist"]
    # more code here
except KeyError as err:
    print("The key '", err, "' does not exist")
    val = "substitute value"
```

You can handle certain exceptions in the try block or all of them. With
Rust we tend to do something like this instead:

```
let value = some_map
    .get("key_does_not_exist")
    .ok_or_else(|| "The key does not exist".into())
    .map_err(|e| {
      eprintln!("{e}");
      e
    })
    .unwrap_or("substitute value");
```

and then we handle the error at the function caller. However, we would
like to handle these and maybe *multiple* of these errors locally. This
is where try/catch would be nice but we just don't have that in Rust
today. If we wanted that we would need to do something like this:

```
let value = match some_map
    .get("key_does_not_exist")
    .ok_or_else(|| "The key does not exist".into()) {
      Ok(val) => val,
      Err(e) => {
        eprintln!("{e}")
        "substitute value"
      }
    };
```

and while this is a contrived case what happens if we have lots of
different errors to handle here that could go wrong? `try/catch` to the
rescue:

```
let value = try_!({
  some_map
    .get("key_does_not_exist")
    .ok_or_else(|| "The key does not exist".into())?
} catch Box<dyn Error> as e{
  eprintln!("{e}")
  "substitute value"
});
```

This looks a bit nicer, we know what kind of error we'll work with and
if we want to handle many local errors in the block we can! This isn't
the solution for *all* errors, but when we do need it, it's really
powerful.